Refactor names, add FileIndexer
Refactors names in the library:
- to distinguish Indexer from Accessor
- to distinguish KEYs from derived NAMEs
- to distinguish external vs internal collections

Adds a FileIndexer, which supports applying fractional indexing
to a component of a filename, defined by an arbitrary user-supplied
name-to-key mapping function.

SqliteIndexer also supports a name-to-key-mapping implicitly,
so long as key ordering predicts name ordering.

added import statement to __init__.py, so that
the top-level module can access the symbols it
export